GKN-108: 1) file placement reorganization, 2) introduce the real usage of file-loader plugin
'index.html', 'img' and 'fonts' moved to the root of the project as of incoherence of the name
'public' (the name 'public' should be used for the folder, that contains directly sharing files by
.e.g. http-server; instead, currently this folder contains the sources for the webpack compiler).

file-loader plugin section in the webpack config currently is not used during build due to direct
links to the assets. In order to fix assets browser chaching problem (new assets with old
names are not updated), all assets should be processed by webpack and these names should be
complemented by its hashes.

As there are many done but not closed issues, the refactoring of all images links to processing
through webpack may cause many git conflicts, so in the scope of the ticket only one image usage
refactored to illustrate the approach. GKN-113 issue had been created for the refactoring of the
rest images.